Implement SegmentedOrderedBinaryTree
Asynchronous createbranchchannel for OrderedBinaryTree
Remove type specialization in createbranchchannels
