+++ test/singlehost.jl
@@ -1,0 +1,8 @@
+using Distributed
+
+const workersused = 8
+addprocs(workersused)
+
+include("tests.jl")
+
+rmprocs(workers())
